First stab at the updated header structure
Updates the header structure and style to match the Twenty Twenty design. This commit also includes the new ”Shortcuts Menu”, which is displayed next to the header toggles on 1000px and up.

The commit also changes all toggles to be button elements instead of links, and changes the aria-pressed attribute to aria-expanded.

The commit does _not_ include accessible drop down menus. I’m guessing we’ll use a JS based solution for that, in which case it might be better to wait for it until the jQuery → vanilla JS update is done.