Set featured media to be half white on all sizes
Due to the previous commit, images can now be thinner than the featured media container on mobile, which means that the pseudo element creating the halfway up white background also needs to be displayed at that size.

Also updated the targeting to include pages and custom post types as well as posts.