Fix animated menus. (#557)
* Fix animated menus.

The previous implementation didn't work, and even if it did, it would
have performance issues due to animating layout-invalidating properties.

This rewrite animates the menu using only opacity and transform, leading
to smooth animations in most devices. The animation is JS-orchestrated,
but entirely CSS-driven. The animation is skipped if a user has
configured their OS or browser to reduce animations.

This change also fixes issues with automatically opening sub-menus on
page load.

Fixes #513.

* Fix animation for menu bottom.

* Fix menu bottom behaviour in IE11.